requirements.txt
streamlit==1.28.0
openai==1.3.0
google-generativeai==0.8.3
chromadb==0.4.18
python-docx==0.8.11
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.6
tiktoken==0.5.1
python-dotenv==1.0.0 